moved parallax to new file to make merges easier
